refactor(db): split monolithic db.ts into domain modules
Break 687-line god file into 7 focused modules under src/lib/db/: types, schema, tasks, sessions, settings, categories, analytics. Barrel index preserves all existing imports via @/lib/db.
