+++ src/lib/db/index.ts
@@ -1,0 +1,38 @@
+export { initDb } from "./schema";
+export { getDb } from "./schema";
+export type { Session, Category, CategoryBreakdown, DayData, WeekSession, WeekSummary } from "./types";
+export {
+  getTasks,
+  addTask,
+  updateTask,
+  deleteTask,
+  toggleTaskArchived,
+  incrementTaskPomos,
+  getTaskTimeToday,
+} from "./tasks";
+export {
+  addSession,
+  startSession,
+  finishSession,
+  abandonSession,
+  getSessions,
+  getTodaySessions,
+  getWeekSessions,
+  getWeekSummary,
+} from "./sessions";
+export { getSetting, setSetting } from "./settings";
+export {
+  getCategories,
+  getCategory,
+  addCategory,
+  updateCategory,
+  deleteCategory,
+} from "./categories";
+export {
+  getCategoryBreakdown,
+  getAllCategoryBreakdown,
+  getWeeklyData,
+  getAllTimeStats,
+  getCurrentStreak,
+  getBestStreak,
+} from "./analytics";
