+++ src/lib/db/settings.ts
@@ -1,0 +1,18 @@
+import { getDb } from "./schema";
+
+export async function getSetting(key: string): Promise<string | null> {
+  const database = await getDb();
+  const rows = await database.select<{ value: string }[]>(
+    "SELECT value FROM settings WHERE key = $1",
+    [key],
+  );
+  return rows.length > 0 ? rows[0].value : null;
+}
+
+export async function setSetting(key: string, value: string): Promise<void> {
+  const database = await getDb();
+  await database.execute(
+    "INSERT INTO settings (key, value) VALUES ($1, $2) ON CONFLICT(key) DO UPDATE SET value = $2",
+    [key, value],
+  );
+}
